refactor(interp): merge arm64 lowering
